back_end/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
pydantic==2.10.2
python-dotenv==1.0.1
pandas==2.2.4
numpy>=1.24.0
faiss-cpu>=1.7.4
